Addded special handling for jsoncpp on Apple, dropped miniball dependency (in favor of CGAL), added explicit cast for JSON vector sizes in some places.

+++ src/run_experiment.cpp
@@ -239,7 +239,7 @@ Json::Value run_task(const moveit::core::RobotModelConstPtr &drone, const Run &r
 	plan_result["run_time"] = run_time;
 	plan_result["start_state"] = (int) run_i;
 	plan_result["scene_name"] = scene->scene_msg->name;
-	plan_result["napples"] = apples.size();
+	plan_result["napples"] = (int) apples.size();
 	plan_result["planner_params"] = planner->parameters();
 	plan_result["planner_name"] = planner->name();
